refactor: user 타입 정리
