+++ src/users/interfaces/user.controller.interface.ts
@@ -1,5 +1,6 @@
 import { FilteredPersonalInfo } from '#auth/types/filtered.types.js';
-import { User, UserPatchDTO } from '#users/user.types.js';
+import { UserPatchDTO } from '#users/types/user.dto.js';
+import { User } from '#users/types/user.types.js';
 
 export interface IUserController {
   patchUser: (body: UserPatchDTO) => Promise<FilteredPersonalInfo<User>>;
