+++ src/users/interfaces/user.repository.interface.ts
@@ -1,7 +1,8 @@
 import { ProviderCreateDTO } from '#auth/types/provider.types.js';
 import { SignUpDTO } from '#auth/types/sign.dto.js';
 import { FindOptions } from '#types/options.type.js';
-import { User, UserUpdateDTO } from '#users/user.types.js';
+import { UserUpdateDTO } from '#users/types/user.dto.js';
+import { User } from '#users/types/user.types.js';
 
 export interface IUserRepository {
   count: () => Promise<number>;
