+++ src/users/interfaces/user.service.interface.ts
@@ -1,6 +1,7 @@
 import { FilteredPersonalInfo } from '#auth/types/filtered.types.js';
 import { FindOptions } from '#types/options.type.js';
-import { User, UserPatchDTO } from '#users/user.types.js';
+import { UserPatchDTO } from '#users/types/user.dto.js';
+import { User } from '#users/types/user.types.js';
 
 export interface IUserService {
   getUsers: (options: FindOptions) => Promise<{ totalCount: number; list: User[] }>;
