+++ src/users/user.repository.ts
@@ -3,7 +3,7 @@ import { SignUpDTO } from '#auth/types/sign.dto.js';
 import { PrismaService } from '#global/prisma.service.js';
 import { FindOptions, SortOrder } from '#types/options.type.js';
 import { IUserRepository } from '#users/interfaces/user.repository.interface.js';
-import { UserUpdateDTO } from '#users/user.types.js';
+import { UserUpdateDTO } from '#users/types/user.dto.js';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
