+++ src/users/user.service.ts
@@ -2,8 +2,8 @@ import { DriverInvalidTokenException } from '#drivers/driver.exception.js';
 import { IStorage, UserType } from '#types/common.types.js';
 import { FindOptions } from '#types/options.type.js';
 import { IUserService } from '#users/interfaces/user.service.interface.js';
+import { UserPatchDTO, UserUpdateDTO } from '#users/types/user.dto.js';
 import { UserRepository } from '#users/user.repository.js';
-import { UserPatchDTO, UserUpdateDTO } from '#users/user.types.js';
 import filterSensitiveData from '#utils/filterSensitiveData.js';
 import { generateS3UploadUrl } from '#utils/S3/generate-s3-upload-url.js';
 import { Injectable } from '@nestjs/common';
