+++ src/utils/filterSensitiveData.ts
@@ -1,6 +1,6 @@
 import { FilteredPersonalInfo } from '#auth/types/filtered.types.js';
 import { Driver } from '#drivers/types/driver.types.js';
-import { User } from '#users/user.types.js';
+import { User } from '#users/types/user.types.js';
 import { generateS3DownloadUrl } from '#utils/S3/generate-s3-download-url.js';
 
 export default async function filterSensitiveData<T extends User | Driver>(data: T): Promise<FilteredPersonalInfo<T>> {
